Conflicts:
	ShopTilYouDrop/src/frontEnd/IramWorkerScreen.java

+++ ShopTilYouDrop/src/frontEnd/HighScoreInterface.java
@@ -5,7 +5,7 @@ import gui6.components.Clickable;
 public interface HighScoreInterface extends Clickable{
 	
 	void setUsername(String user);
-	void setScore(int score);
+	void setScore(double score);
 	int getScore();
 	String getUsername();
 }
